+++ src/r-wrappers.h
@@ -82,7 +82,7 @@ SEXP SXP_get_worst_GEBV(SEXP exd, SEXP s_eff_set);
 SEXP SXP_find_crossovers(SEXP exd, SEXP s_parentFile, SEXP s_outFile, SEXP s_windowSize, SEXP s_certainty);
 SEXP SXP_send_map(SEXP exd);
 
-SEXP SXP_get_groups(SEXP exd, SEXP s_maxGroups);
+SEXP SXP_get_groups(SEXP exd);
 SEXP SXP_get_group_data(SEXP exd, SEXP s_group, SEXP s_whatData, SEXP s_which);
 SEXP SXP_get_group_gene_data(SEXP exd, SEXP s_group, SEXP s_countAllele);